split the linear algebra into modules
