+++ src/asgard_kronmult.hpp
@@ -2,7 +2,6 @@
 
 #include "asgard_block_matrix.hpp"
 #include "asgard_indexset.hpp"
-#include "asgard_kronmult_common.hpp"
 
 namespace asgard::kronmult
 {
