+++ src/asgard_resource_groups.hpp
@@ -218,64 +218,4 @@ private:
   #endif
 };
 
-#ifdef ASGARD_USE_MPI
-/*!
- * \brief Optional call to library initialization
- *
- * The ASGarD library itself does not require initialization; however, ASGarD may be using
- * components that require initialization, e.g., MPI.
- * The final executable can initialize MPI in one of two ways:
- * \code
- *   #include "asgard.hpp"
- *
- *   int main(int argc, char **argv) {
- *     asgard::libasgard_init(argc, argv);
- *     ...
- *     asgard::libasgard_finish();
- * \endcode
- * The methods will call the proper initialization methods, if MPI has been enabled,
- * otherwise the methods will do nothing.
- *
- * Alternatively, MPI_Init can be called directly
- * \code
- *   #include "asgard.hpp"
- *
- *   int main(int argc, char **argv) {
- *     #ifdef ASGARD_USE_MPI
- *     MPI_Init(&argc, &argv);
- *     #endif
- *     ...
- *     #ifdef ASGARD_USE_MPI
- *     MPI_Finalize();
- *     #endif
- * \endcode
- *
- * Naturally, the #ifdef directives can be omitted or the initialization can be skipped
- * altogether, if the given PDE definition always or never uses MPI.
- */
-inline void libasgard_init(int &argc, char **&argv) {
-  MPI_Init(&argc, &argv);
-}
-//! finalization of the library, see asgard::libasgard_init
-inline void libasgard_finish() {
-  MPI_Finalize();
-}
-//! RAII style call to init/finish
-struct libasgard_runtime {
-  //! calls libasgard_init
-  libasgard_runtime(int &argc, char **&argv) {
-    libasgard_init(argc, argv);
-  }
-  //! calls libasgard_finish
-  ~libasgard_runtime() { libasgard_finish(); }
-};
-#else
-inline void libasgard_init(int &, char **&) {}
-inline void libasgard_finish() {}
-struct libasgard_runtime {
-  //! does nothing
-  libasgard_runtime(int &, char **&) {}
-};
-#endif
-
 }
